clean up raw disc reading code a bit and add game title to window
also refactor the PhysicalMedium trait to only contain the methods that are actually needed by Memory<PhysicalMedium>
