+++ segacd-core/src/cdrom.rs
@@ -1,4 +1,5 @@
+pub mod cdtime;
 pub mod cue;
 pub mod reader;
 
-const BYTES_PER_SECTOR: u64 = 2352;
+pub const BYTES_PER_SECTOR: u64 = 2352;
